Corrected onclick behavior and other minor fixes
Changed the onclick function attached to the markers on the map:
previously it would open the first popup that had the same lat as the clicked point, causing some error when two points had the same lat.

Now, check for both coordinates to select the appropriate popup

Also change an incorrect use of var i in plotMarker, which should have been (and is now) data_index, but somehow was functional.

Changed behaviour of search bar: when clicking on a result, the map does not zoom that much in on the selected point (used to be 17, now 15), which should avoid tile not loading. Also, had the spider and the pop-up show up when using the search bar, but the results are not 100% satisfactory (see comments in code, in activateSearchResults()